Draft push logic.
When a client detects that it is ahead of a remote server it attempts to
push the latest version of the vault to the remote server. Requires work
to store payloads and send a PATCH request with a batch of payloads
rather than a brute force PUT of the entire vault buffer.

Also begin work to support offline first logic on the client.