Handle conflicts (#10)
* Draft conflict handling.

Testing with a simple read operation but the client has a bug where it
is not updating the secret data immediately. The remote vault buffer has
been loaded so this looks like a react caching or invalidation issue.

* Basic conflict resolution.

Pull from the server when the local copy is behind and replay the last
operation.

* Support save vault operation.

So that a local vault that is ahead of the remote may be pushed when
resolving conflicts.

* Update to new binary traits.

* Check vault summary when accepting byte buffer.

To verify a buffer payload looks like a valid vault.

* Draft push logic.

When a client detects that it is ahead of a remote server it attempts to
push the latest version of the vault to the remote server. Requires work
to store payloads and send a PATCH request with a batch of payloads
rather than a brute force PUT of the entire vault buffer.

Also begin work to support offline first logic on the client.

* Draft logic to queue unsaved changes client side.

Stored as a changeSet batch keyed by vault UUID. Requires work to send
PATCH requests to the server for processing.

* Add Patch type.

* Prepare to send patch changeset to server.

* Add network guard for patchVault() call.

* Update decode() implementation to use a reference.

Now possible since the updates to serde_binary use a SliceStream for
decoding.

* Preparing patch server handler.

* Tidy conflict handling types and logic.

* Convert Payload to ServerEvent.

* Draft patch handler logic.

* Tidy Cow handling in payload encode.

* Update serde-binary dependency.