Improve implementation for the mv command.
So we get the encrypted data directly from the vault, call to the
gatekeeper to decrypt the meta data and mutate that before re-encrypting
and calling update() using the original secret AEAD pack.

This prevents the need for decrypting the secret data when we only need
to operate on the meta data to rename a secret and saves adding an
additional function to the VaultAccess trait.