Improve handling of change notifications.
So that events are grouped together in a single message; when a patch
containing multiple sync events is applied this will only create a
single notification rather than previously it would send several
notifications which could have caused clients receiving the
notifications to make lots of rapid requests to the server to update
their state.

Now clients can inspect the events in the change notification to
determine what action to take including reducing updates down to a
single server request when possible.

Closes #37.