Support changing encryption passphrase. (#51)
* Prepare logic for changing encryption passphrase.

* Add unit test for change password logic.

* Prepare backend for vault overwrite.

Implement a set_vault() function for the change password scenario.

* Prepare server handler.

* Prepare client disc cache for change passphrase.

* Draft change password implementation.

Commented out some integrity checking code that is failing at the
moment.

Need to create an integration test and for this very important flow.

* Improving integration tests.

* Add delete secret to integration test.

* Add remove vault to integration test.

* Improve integration tests.

* Update code coverage tasks.

* Improve integration test coverage.

* Find fix for the unknown issuer SSL bug.

When running locally and trying to connect to the changes feed reqwest
was throwing an "Unknown Issuer" SSL error. This was due to disabling
`native-tls` with `default-features = false`; but we require switching
off `default-features` to omit openssl on `aarch64-unknown-linux-gnu`.

This should fix the local development workflow and continue to build
correctly for ARM linux.

* Update handling of reqwest dependency.

* Restore release workflow.

* Prepare to assert on change notifications.

* Assert on change notifications in integration test.

* Prepare for multiple clients in integration tests.

* Prepare for multiple integration tests.

* Draft change password integration test.

* Fix sort order issue with WAL events.

When splitting records on the server from the vault sent by the client
the order of the `CreateSecret` events did not match the order the
client had which caused the root hashes to be different.

* Fix last commit logic using apply() in WAL files.

As the buffer had not been flushed to disc the last_commit() blanket
implementation could not find a last commit correctly so ensure we
compute it when iterating in apply().