Verify message identity integrity.
The RPC client now sets a bearer signature for every method call and the
server verifies that the address for the bearer signature matches the
address assigned to the session.

Hardens the identity check by not relying solely on the session
identifier to establish the identity of the caller.

The input bytes for the signature are a digest of the session secret key salt (which is also the session challenge) concatenated with the bytes for the message nonce.