Improve status code handling.
So that the service handler will use the inner status code as the actual
HTTP status code for the response with some exceptions.

For example, NOT_MODIFIED is mapped to OK so that the HTTP client will
read the data from the response body.

The RPC client then returns the *inner* status code included in the
ResponseMessage so client code can handle the NOT_MODIFIED scenario.