Remove indirection on encryption passphrase.
By storing a different encryption passphrase in the login vault it would
make a user vulnerable to a denial of service attack.

If the login vault was deleted but the user still had files for the
other vaults containing their user data they would not be able to access
the secrets. By sharing the master vault password across all vaults we
remove this attack vector.