Refactor SearchIndex to Arc<RwLock>.
To support virtual folders in the UI which will filter by type
identifier.

Previously, the assumption was that searching was performed on a single
current vault which is why the Gatekeeper owned the SearchIndex.

Now we have a requirement to search across multiple vaults and filter by
type identifier the SearchIndex must be injected into the Gatekeeper.