Prepare new migration archive layout.
Use folders and files for each vault/secret and extract file secret
buffers into separate files.

This prevents us from needing to base64 encode the buffers for file
secrets and attachments which would add overhead to the webassembly
bindings and make the JSON bloated.

Requires work to pre-compute the checksums for files/attachments.