Buffer in file storage logic.
This one is tricky as AGE requires the sync Read/Write traits and we
don't want to use std::fs for webassembly support so the only workaround
for now is to buffer the entire file into memory when
encrypting/decrypting.

This is far from ideal but unless AGE ends up supporting AsyncRead and
AsyncWrite there is not way around this for now.