Fix bug with folder import logic, sync test now passes.
When calling create_cache_entry() in the local provider it automatically
appends the create vault event to the event log, however when importing
and overwriting a vault this would cause the create vault event to
appear twice in the log file.

Now we truncate() the event log before appending the create vault event.