Deleting account not longer moves to a trash folder.
It was a bad idea to try to preserve deleted accounts as we would need
to provide a lot of functionality around recovering accounts from the
trash and managing contents of the trash folder and it may well be the
case that people actually want to completely delete the secrets.

For example, if somebody is about to become under duress deletion should
mean deletion.