Workaround for bug with duplicate secrets.
When pairing the paired account was receiving a WriteEvent::CreateVault()
event containing a vault with some secrets which would result in
duplicates as the secrets would be inherited from the create vault event
and also inserted from the subsequent events.

The workaround for now is to truncate() the vault before inserting the
entries from the events but we need better handling of the CreateVault
variant so it never contains additional secrets.