Use pending directory for LOCAL folder stubs.
So that other client devices can respond to AccountEvent::CreateFolder
when a LOCAL folder is created on another device by creating stub files
in the pending directory.

When a merge is attempted on a folder we first try to promote the folder
in case it is pending.

This allows the NO_SYNC flag to backup folders to remote server(s) but
not share the contents with other devices until the NO_SYNC flag is
removed.