Refactor to prefer LocalRequest and LocalResponse.
Deletes the now obsolete IpcRequest and IpcResponse types and removes the
logic for delegating from a service to an app.

The new design is that app integrations must sync account data using the
local HTTP server and the standard sync protocol and either store the
data in-memory or cache elsewhere on disc. Users must authenticate to
account(s) in the app integration to give permission for the app to
access the data.