Fix type definition.
When serde sees a tuple enum variant with two entries but the last one
is set to skip then it encodes as an array of one item, however in the
case of WriteEvent::DeleteSecret it's a single item and the containing
array is skipped. I suspect because it is serialized using the new type
logic.