Fix logic for watching for folder create.
Watch the identity event log for changes rather than the account events
because changes to the account events would fire earlier than changes
to the identity folder making it hard to reconcile when the new folder
password would have been saved in the identity folder.

By watching the identity folder event log instead we know that we can
reload the identity folder to access the new folder password which is
required to make watching work well when new secrets are created within
a new folder.