Use in-memory journaling for database upgrade.
We need a complete database file during the database upgrade
otherwise some database integration tests will fail sporadically.

We could VACUUM at the end of the database upgrade to create a complete
database file which works but instead I've opted to use an in-memory
journal mode which also fixes the problem.

Relying on the less-reliable in-memory journaling mode for database
upgrade is considered OK as the final database file is only moved into
place once the entire upgrade is completed and has been verified.