Support upgrading v1 backup archives.
Modify the v2 backup archive so it has an explicit version 2 (it was
exported before we explicitly set the version).

Add the bypass_status_check option to the UpgradeOptions just to
support upgrading v1 backup archives which do not include all the
account event logs. Because they don't have all the required event logs
it's not possible to call sync_status() to perform an integrity check
between the source filesystem account and the target database backend.